Updated trail to be a smart pointer (shared) instead of a raw pointer.
Trails will now auto deallocate when out of scope. May, want to make it a unique_ptr to prevent any weird copies in future development. However, this would require some form of custom deep copying if functions stay as they are. That is, when returning comet on line 195 it is implicitly making a copy which is not allowed with a unique pointer.

With the shared, within that function we make the shared and it will have a ref count of 1, it hits the return and briefly has a count of 2 but then the original comet object goes out of scope so the count goes back to 1.